# Extending KEVM configuration with `<wordStackSize>`