--- a/media/201908-trufflecon/add-spec.k
+++ b/media/201908-trufflecon/add-spec.k
@@ -6,11 +6,11 @@ module ADD-SPEC
     rule <k> #next [ ADD ] => . ... </k>
          <schedule> SCHEDULE </schedule>
          <wordStack> W0 : W1 : WS => chop(W0 +Int W1) : WS </wordStack>
-         <wordStackSize> WSSize => WSSize -Int 1 </wordStackSize>
+         <wordStackSize> WSSIZE => WSSIZE -Int 1 </wordStackSize>
          <pc> PCOUNT => PCOUNT +Int 1 </pc>
          <gas> G => G -Int Gverylow < SCHEDULE > </gas>
       requires G >=Int Gverylow < SCHEDULE >
-       andBool notBool ( #stackUnderflow ( WSSize , ADD )
-                  orBool #stackOverflow  ( WSSize , ADD )
+       andBool notBool ( #stackUnderflow ( WSSIZE , ADD )
+                  orBool #stackOverflow  ( WSSIZE , ADD )
                        )
 endmodule