--- a/optimizer/optimize-spec.k.tmpl
+++ b/optimizer/optimize-spec.k.tmpl
@@ -13,12 +13,13 @@ module {MODULE}-SPEC
     claim <k> (#next [ {OPCODE} ] => #halt) ~> #execute ... </k>
           <schedule> SCHED </schedule>
           <program>  PGM   </program>
-          <pc>         PCOUNT       => PCOUNT +Int #widthOp({OPCODE}) </pc>
-          <wordStack>  {WORD_STACK} => ?_                             </wordStack>
-          <localMem>   LM           => ?_                             </localMem>
-          <gas>        #gas(GAVAIL) => ?_                             </gas>
-          <memoryUsed> MU           => ?_                             </memoryUsed>
-      requires #sizeWordStack({WORD_STACK}) +Int #stackDelta({OPCODE}) <=Int 1024
+          <pc>             PCOUNT            => PCOUNT +Int #widthOp({OPCODE}) </pc>
+          <wordStack>      {WORD_STACK}      => ?_                             </wordStack>
+          <wordStackSize>  {WORD_STACK_SIZE} => ?_                             </wordStackSize>
+          <localMem>       LM                => ?_                             </localMem>
+          <gas>            #gas(GAVAIL)      => ?_                             </gas>
+          <memoryUsed>     MU                => ?_                             </memoryUsed>
+      requires {WORD_STACK_SIZE} +Int #stackDelta({OPCODE}) <=Int 1024
        {EXTRA_REQUIRES}
 
 endmodule