--- a/tests/failing/ContractCreationSpam_d0g0v0.json.expected
+++ b/tests/failing/ContractCreationSpam_d0g0v0.json.expected
@@ -56,6 +56,9 @@
         <wordStack>
           .WordStack
         </wordStack>
+        <wordStackSize>
+          0
+        </wordStackSize>
         <localMem>
           b""
         </localMem>