--- a/tests/failing/static_callcodecallcodecall_110_OOGMAfter_2_d0g0v0.json.expected
+++ b/tests/failing/static_callcodecallcodecall_110_OOGMAfter_2_d0g0v0.json.expected
@@ -56,6 +56,9 @@
         <wordStack>
           .WordStack
         </wordStack>
+        <wordStackSize>
+          0
+        </wordStackSize>
         <localMem>
           b""
         </localMem>