--- a/tests/specs/benchmarks/bytes00-spec.k
+++ b/tests/specs/benchmarks/bytes00-spec.k
@@ -23,7 +23,8 @@ module BYTES00-SPEC
           <caller> MSG_SENDER </caller> // msg.sender
           <callData> #abiCallData("execute", #bytes(#bufStrict(DATA_LEN, DATA))) </callData> // msg.data
           <callValue> 0 </callValue> // msg.value
-          <wordStack> .WordStack => ?_ </wordStack>
+          <wordStack> .WordStack => ?WS </wordStack>
+          <wordStackSize> 0 => #sizeWordStack(?WS) </wordStackSize>
           <localMem> .Bytes => ?_ </localMem>
           <pc> 0 => ?_ </pc>
           <gas> #gas(_VGAS) => ?_ </gas>
@@ -135,4 +136,3 @@ _
      andBool #range(0 <= DATA < (2 ^Int (DATA_LEN *Int 8)))
 
 endmodule
-