--- a/tests/specs/benchmarks/ecrecoverloop02-sig1-invalid-spec.k
+++ b/tests/specs/benchmarks/ecrecoverloop02-sig1-invalid-spec.k
@@ -27,7 +27,8 @@ module ECRECOVERLOOP02-SIG1-INVALID-SPEC
 #bytes32(SIGR0), #bytes32(SIGR1),
 #bytes32(SIGS0), #bytes32(SIGS1)) </callData> // msg.data
           <callValue> 0 </callValue> // msg.value
-          <wordStack> .WordStack => ?_ </wordStack>
+          <wordStack> .WordStack => ?WS </wordStack>
+          <wordStackSize> 0 => #sizeWordStack(?WS) </wordStackSize>
           <localMem> .Bytes => ?_ </localMem>
           <pc> 0 => ?_ </pc>
           <gas> #gas(_VGAS) => ?_ </gas>
@@ -151,4 +152,3 @@ _
      andBool         #ecrecEmpty( #bufStrict(32, HASH), #bufStrict(32, SIGV1), #bufStrict(32, SIGR1), #bufStrict(32, SIGS1) )
 
 endmodule
-