--- a/tests/specs/benchmarks/encode-keccak00-spec.k
+++ b/tests/specs/benchmarks/encode-keccak00-spec.k
@@ -23,7 +23,8 @@ module ENCODE-KECCAK00-SPEC
           <caller> MSG_SENDER </caller> // msg.sender
           <callData> #abiCallData("execute", #bytes32(A0), #bytes32(A1)) </callData> // msg.data
           <callValue> 0 </callValue> // msg.value
-          <wordStack> .WordStack => ?_ </wordStack>
+          <wordStack> .WordStack => ?WS </wordStack>
+          <wordStackSize> 0 => #sizeWordStack(?WS) </wordStackSize>
           <localMem> .Bytes => ?_ </localMem>
           <pc> 0 => ?_ </pc>
           <gas> #gas(_VGAS) => ?_ </gas>
@@ -135,4 +136,3 @@ _
      andBool #rangeUInt(256, A1)
 
 endmodule
-