… kevm-pyk/src/kevm_pyk/kproj/evm-semantics/evm.md
… kevm-pyk/src/kevm_pyk/kproj/evm-semantics/optimizations.md
… kevm-pyk/src/kevm_pyk/kproj/evm-semantics/state-utils.md
… tests/specs/benchmarks/requires01-a0gt0-spec.k
… tests/specs/benchmarks/requires01-a0le0-spec.k
… tests/specs/benchmarks/staticarray00-spec.k
… tests/specs/benchmarks/staticloop00-a0lt10-spec.k
… tests/specs/benchmarks/storagevar00-spec.k
… tests/specs/benchmarks/storagevar01-spec.k
… tests/specs/benchmarks/storagevar02-nooverflow-spec.k
… tests/specs/benchmarks/storagevar02-overflow-spec.k
… tests/specs/benchmarks/storagevar03-spec.k
… tests/specs/benchmarks/structarg00-spec.k
… tests/specs/benchmarks/structarg01-spec.k
… tests/specs/erc20/ds/allowance-spec.k
… tests/specs/erc20/ds/approve-failure-spec.k
… tests/specs/erc20/ds/approve-success-spec.k
… tests/specs/erc20/ds/balanceOf-spec.k
… tests/specs/erc20/ds/totalSupply-spec.k
… tests/specs/erc20/ds/transfer-failure-1-a-spec.k
… tests/specs/erc20/ds/transfer-failure-1-b-spec.k
… tests/specs/erc20/ds/transfer-failure-1-c-spec.k
… tests/specs/erc20/ds/transfer-failure-2-a-spec.k
… tests/specs/erc20/ds/transfer-failure-2-b-spec.k
… tests/specs/erc20/ds/transfer-success-1-spec.k
… tests/specs/erc20/ds/transfer-success-2-spec.k
… tests/specs/erc20/ds/transferFrom-failure-1-a-spec.k
… tests/specs/erc20/ds/transferFrom-failure-1-b-spec.k
… tests/specs/erc20/ds/transferFrom-failure-1-c-spec.k
… tests/specs/erc20/ds/transferFrom-failure-1-d-spec.k