This PR introduces a `<wordStackSize>` cell into the EVM configuration, which holds the size of the word stack. This should remove the need to continuously compute the size of the word stack, which is both recursive and cannot be done using the LLVM backend if any of the elements of the word stack are symbolic.

What needs to be done carefully in this change is to preserve the invariant that having
```
<wordStack> WS </wordStack>
<wordStackSize> WSSize </wordStackSize>
```
means that `WSSize ==Int #sizeWordStack(WS)`.