@PetarMax 2024-09-23
On one of our engagement test suites, the performance is as follows (with parallel factor 6):

|     | wssize | master | speedup |
|----:|-------:|-------:|--------:|
|  T1 |    170 |    175 | -2.86%  |
|  T2 |    133 |    139 | -4.32%  |
|  T3 |    104 |    113 | -7.96%  |
|  T4 |    316 |    342 | -7.60%  |
|  T5 |   1573 |   1699 | -7.42%  |
|  T6 |    272 |    339 | -19.76% |
|  T7 |    165 |    200 | -17.50% |
|  T8 |    104 |    111 | -6.31%  |
|  T9 |   4744 |   4901 | -3.20%  |
| T10 |    118 |    128 | -7.81%  |
| T11 |    927 |    998 | -7.11%  |

so we can see that there is a speed-up throughout, with varying extent.