@PetarMax 2024-09-23
On another engagement-related test suite, the times are as follows:

|     | wssize | master | speedup |
|----:|-------:|-------:|--------:|
|  T1 |     42 |     42 | 0.00%   |
|  T2 |     34 |     34 | 0.00%   |
|  T3 |     39 |     45 | -13.33% |
|  T4 |     41 |     47 | -12.77% |
|  T5 |    262 |    365 | -28.22% |
|  T6 |    477 |    581 | -17.90% |
|  T7 |    131 |    180 | -27.22% |

with a higher speed-up on average.

I think that this depends on just how much stack manipulation is happening.